+++ src/ItemTest.php
@@ -358,6 +358,8 @@ final class ItemTest extends StructuredFieldTestCase
         self::assertNotSame($instance1, $instance3);
         self::assertEquals($instance1->value(), $instance3->value());
         self::assertSame($instance1, $instance4);
+        self::assertTrue($instance1->parameter('a'));
+        self::assertNull($instance5->parameter('a'));
         self::assertTrue($instance5->parameters()->hasNoMembers());
         self::assertTrue($instance6->parameters()->hasNoMembers());
     }